Move to api folder:
- SignupAPI
- ProfileAPI
- MyProfileAPI
