--- app/forms.py
@@ -1,24 +1,0 @@
-from django import forms
-
-from django.contrib.auth.models import User
-
-class SignUpForm(forms.Form):
-    first_name = forms.CharField(max_length=100, required=True)
-    last_name = forms.CharField(max_length=100, required=True)
-    email = forms.EmailField(required=True)
-    password = forms.CharField(widget=forms.PasswordInput(), required=True)
-    password_confirm = forms.CharField(widget=forms.PasswordInput(), required=True)
-    birthdate = forms.DateField()
-
-    def clean(self):
-        cleaned_data = super().clean()
-        password = cleaned_data.get("password")
-        password_confirm = cleaned_data.get("password_confirm")
-
-        if password != password_confirm:
-            self.add_error('password_confirm', "Password does not match")
-
-        if User.objects.filter(username=cleaned_data.get('email')).first():
-            self.add_error('email', "Email is already used")
-
-        return cleaned_data
